Sysinv tox updates. Prepare for bandit reports and test reports
The old sysinv py27 target did not use testrepository so its results
were not reported in jenkins. Now it will.
The bandit target will now output to an xml file which can be parsed and
reported in jenkins.

Note:  running tox for sysinv will reports warnings about missing
fmClientUI and missing ceph. this is acceptable as neither of those are
installed in the tox environment, and are side effects of using fm-api

Change-Id: I48c74e9a62cd6fd32ceb71e1994c55fa7e8c42cf