Package in helm Create helm package
There is no upstream RPM for the helm binary, so we need to make one.

The upstream helm source/binary tarballs are available at:
https://github.com/kubernetes/helm/releases

Add the docker image for "tiller" to be used by helm, and ensure that
it gets initialized properly on controller node startup.

Including the docker image in the RPM is only a temporary solution
until we figure out the best way to handle docker images.

The changes to kubernetes.pp are to ensure that the nested dependencies
are completed before the helm manifest tries to run commands that depend
on docker/kubectl being functional.

Story: 2002876
Task: 22831
Depends-On: https://review.openstack.org/587390

Change-Id: Icf7e73047aecc933bf3451b20b6159997dfbea99
Signed-off-by: Jack Ding <jack.ding@windriver.com>