Provide a way to enable/disable NTP service
Modify NTP puppet manifest to allow enabling and disabling this service.
This is required in case user would like to enable PTP service instead.
As a result, NTP daemon must be stopped on compute/storage nodes.
It is still running on controller nodes, but all the servers are wiped.
So, system date cannot be updated by NTP on controller nodes, but
it is still possible for compute nodes to do an initial time sync.
Also, PMON script is moved here from MTCE to follow enabled/disabled path.

Change-Id: I0831487fb14de80edec0ee8df5cc6f23dfb6bae8
Story: 2002935
Task: 24646
Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>