Fix sysinv pep8 failure of type F841
F841: local variable 'db_values' is assigned to but never used.

The unused variable is located within a try block in helm_charts.py
This line is required as the code checks to see if it throws an
exception.

A comment was added to make the intention of this line clearer.

Story: 2002888
Task: 23097

Change-Id: Ic9d65761db64ed44ea48de2de70d2da8ef21aa10
Signed-off-by: Mathieu Robinson <mathieu.robinson@windriver.com>