Set the type of endpoint to internalURL on all types of nodes
Missing ceilometer measures cpu, cpu_util and vcpu_util with https
enabled system due to public URL is currently used in ceilometer
as the type of endpoint for communication with openStack services.
This update is to set the type of endpoint to internalURL through
puppet for all types of nodes.

Closes-Bug: 1790238

Change-Id: I19f0e354a22b089f140ed8e762f3db68f6bc4ddb
Signed-off-by: Angie Wang <angie.wang@windriver.com>