Simplify disk space allocation
Allocating of space on the root disk was done in kickstart, combined
with config_controller. Various checks for disk and volume group
space made it difficult to add new filesystem partitions.

To simplify this, all checks are now merged under disk size checks.

Kickstart files now create the partitions of the same size for
rootfs and log partitions, no matter the disk size, with the
only variable size being the cgts-vg size.

Thus, any future changes will only need to consider only this
size for ensuring enough space is present.

The limit between small and big disks is 240GB, with a minimum
disk size of 120GB.

Change-Id: I0ac54d5b2ba72af9505881c5849bc4e76540b3f4
Closes-bug: 1791170
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>