tox flake8, pep8 checks only .py files
By default flake8 checks only *.py files. There are some
python scripts that don't have .py extensions and therefore
are skipped by flake8:

  sysinv/cmd/manage-partitions
  sysinv/cmd/query_pci_id
  sysinv/cluster/cluster_services_dump

A solution would be to run flake8 on all files identified
by 'file' as 'Python script' but because that's time consuming
and the number of scripts is low and unlikely to change
we're hard-coding extra files to be checked in flake8's
command line.

The command to scan the project for not ending in .py
python scripts and update tox.ini file is also added
as a comment.

Change-Id: I902bcbbcd59a60660423d1223a5d5aef08ae6d9f
Closes-bug: #1791350