Fix exceptions from nova/neutron client access
When the --kubernetes option is used with config_controller, native
services will not be running. In this scenario, nova and neutron APIs
will not be running if the containerized openstack services are not
installed.

Make sure that attempts to access select nova and neutron client calls
from the conductor handle the connection exception. This is done so that
logs are clean and remote RPC calls from the agent, which would also
raise an exception, do not block normal agent audit operations when
attempting to assign storage host aggregates.

This does not handle every client call, but is targeted to exceptions
that are currently being thrown during initialization.

Change-Id: Ifc904208bd3e30c73a0a5e0740c00fb5a02982d8
Story: 2002876
Task: 26866
Signed-off-by: Robert Church <robert.church@windriver.com>