stx-config: API ref doc content added.
I added the api-ref-sysinv-v1-config.rst file, which is content
for the API reference manual to the api-ref/source dir. This
represents the converted old-style files to the newer OpenStack
supported RST files.

I updated the index.rst file to include the new .rst file so
that the api-ref document can build.

Change-Id: Ibcac482b724ceeac51fe3d4471022f4c6f31ee2d
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>