Networks now assigned to interfaces when interface profile is applied
This commit adds 'networks' to the list of fields to be copied in/out
of interface profiles. Now, when an interface profile is applied to
a node, the networks will also be assigned to the appropriate interfaces.

In addition, the command 'system interface-network-list' now requires a
host ID/name instead of the field being optional.

Closes-Bug: 1794546

Change-Id: I739c355c06f6adbea4444fabe45a5032827f6bf5
Signed-off-by: Patrick Bonnell <patrick.bonnell@windriver.com>