Fix host-label duplication
The host label key and value are changed to be stored in
separate fields in the database. A unique constraint
between the host and the label key is added.
The validation of the label key and value is performed the
kubernetes client api instead of replicating the checking
in the host-label api. The error message from the kubernetes
client api is passed to the host-label api.

Story: 2002845
Task: 24936

Change-Id: I7e318d5e58602d33652d49452e35d99322565148