Containerization, support 2 keystones in sysinv
Support bare metal and pod based keystone in sysinv.  An extra config
section is added to sysinv.conf with duplicated keystone parameters,
and the openstack region name parameters are moved to this new openstack
keystone section.  The original keystone section is used for platform
service authentication.  Authentication parameters are set to the same
values in the 2 config sections for now.

Change-Id: I7a53dd5a2dc52213e0f1e0cc748649a33f0f9f40
Story: 2002876
Task: 26926
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>