Standby controller filesystem sizes check
While installing a system with two controllers, you can assign new
PVs to the cgts-vg volume group and resize the filesystem on
controller-0, before provisioning controller-1.

If these new sizes are above the default provisioned space for
cgts-vg on controller-1, the unlock is allowed, but the node
goes into a reboot loop due to not havign enough space to
assign to all the partitions.

Now, before unlocking the standby controller for the first time
we check if the provisioned space on the node is equal or larger
than the used space on the active controller and if it is not
reject the unlock.

Change-Id: I3fce3430abbb81d08272f35915cc50c761754733
Closes-bug: 1797108
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>