Horizon Helm Chart Override Generation
This commit adds horizon's required HELM chart overrides to sysinv.
The usage of these new overrides are used by the local_settings
template which is overridden in the armada manifest.  This approach
was taken because fitting a multi-hundred line override into sysinv
would be prety messy.

Still TODO:
- Sorting out how this will work in Region and
  Distributed Cloud deployments
- Enabling outside access to containerized horizon

Depends-On: 610979
Change-Id: I0df2e1441c9a44a6e3451e8cdefbbd28367107f4
Story: 2003909
Task: 27080
Signed-off-by: Tyler Smith <tyler.smith@windriver.com>