Lock storage nodes during system restore
Currently, as part of system restore, maintenance will lock all the
nodes except for controller-0. When locking storage nodes if the
replication will be lost, we query ceph pools to see if they are
empty. But during restore, ceph cluster is not up yet; therefore
query ceph pools will fail and it will block locking storage nodes.

Solution: In the case of the restore, if it is a force-lock, we
should skip the ceph pool query and just lock the storage node.

Change-Id: I0501b44ebe635f36d7401437bfe289de7ee9fc73
Closes-Bug: 1798197
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>