Support System B/R without Reinstalling Storage Nodes
This code change is to support system B/R without reinstalling
storage nodes.

The config_controller command is modified as follows:

  "config_controller --restore-system <name>" is replaced with
  following two options:

  - config_controller --restore-system --include-storage-reinstall
    <name>
  - config_controller --restore-system --exclude-storage-reinstall
    <name>

Add a new command "tidy_storage_post_restore" to be run after
the restore. It scans Cinder/Glance and rbd backend for any
discrepancy between the two and generates a user action log file.

Story: 2004184
Task: 27672

This review is being marked as WIP and workflow set to -1. This code
change has been partially tested on a virtual box as well as hardware
lab. I am requesting feedback from reviewers familiar with B/R to
ensure this approach is acceptable before completing the testing and
updating this commit.

Change-Id: I5e64fde70b977ea4bc3a5927bcbd852a393caec2
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>