Enable StarlingX helm charts for stx-openstack app
Use an application RPM to build and collect application artifacts.

This follows the same mechanism that is used to build the openstack-helm
and openstack-helm-infra charts. The custom StarlingX charts are built
and the application specific armada manifests are included in the RPM.

This application RPM will be used to produce some of the required build
artifacts (custom helm tarballs + armada manifests) but is not intended
to be installed on the system. These artifacts are extracted later for
proper application packaging with additional required metadata.

Changes also include:
 - Update the README describing the new layouts for applications.
 - Move the nova-api-proxy chart from the common area to be included as
   a chart specific to the stx-openstack app.

Change-Id: I1140760d56035249324519be93bb913e18f394d9
Story: 2004005
Task: 27801
Depends-On: I57c5ec5f3565e9e585f0935af745e495699aa28c
Signed-off-by: Robert Church <robert.church@windriver.com>