Add Aodh and Panko to the armada manifests
Aodh and panko charts are added to the openstack-telemetry chart group,
but openstack-telemetry group still needs to be temporarily disabled
as gnocchi requires ceph storage which is not supported in 2 nodes
confiugration yet.

In the meantime, user will need to manually enable openstack-telemetry
in the armada-manifest chart groups.

Change-Id: Ia8f4ad7ddbe85a54b6774d3df1666d3fc858c8ca
Depends-On: https://review.openstack.org/#/c/618006/
Story: 2003909
Task: 27083
Signed-off-by: Angie Wang <angie.wang@windriver.com>