Cinder-volumes Volume Group Resize Procedure Failed
When the disk partition for the cinder-volumes VG is increased
on the newly standby controller after controller swact, drbd sync
starts before cinder-volumes VG is extended. When drbd sync starts,
the drbd is in "SyncSource" state. But the code waits for drbd to
be in "Connected" state. If that doesn't happen within 5 seconds,
it will fail and won't continue to extend cinder-volumes VG.

The solution is to allow cinder-volumes VG extension while drbd is
in either "Connected" or "SyncSource" state.

Change-Id: Ib164ce5fb433789d0806ac8d4faae9d675695b65
Closes-bug: 1803772
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>