Enable communication from nova-api-proxy to VIM
Prior to kubernetes integration, the nova-api-proxy
communicated with the VIM using port 30003 on the loopback
address. This worked fine with both the nova-api-proxy and
the VIM were running on bare metal.

When nova-api-proxy is running is running in a pod, it
cannot use the loopback address to communicate with the VIM,
since the loopback address is local to the pod. The
nova-api-proxy pod will instead use the floating management
address (which is present on all systems) to communicate with
the VIM. The problem is that on AIO simplex systems, the
floating management address is also in the loopback subnet.
The fix for this is to use a non-loopback address for
management on AIO simplex systems. This required changes to
configuration scripts/utilities to allow the user to
specify the management subnet on AIO simplex systems,
instead of hardcoding it to a loopback address.

Change-Id: I26c15657471f8214e628c6d43eaab07eb8e744ec
Story: 2003910
Task: 28061
Depends-on: Ib7c08bfc1c661a9fb347308e55cc806b4dd129ad
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>