Display available space in logical volume group
Currently there is no easy way for user to know the available
space in each logical volume group. This commit adds a new
column/row to show the size of available space in
host-lvg-list and host-lvg-show command output.

This commit also fixes a bug in cgts-vg available space
calculation and removes unused code.

Change-Id: Ic94a3544a9d02f67c592e07048b95135c2b350ed
Closes-Bug: 1804711
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>