Glance overrides - increase replicas count
If a CEPH backend is enabled and the glance storage
is rbd, then set the number of replicas to the number
of controllers.

If there is no CEPH backend present and the glance
storage is pvc, then the replicas must be set to 1
as the glance chart doesn't create two PVCs, as needed.

Change-Id: I293b7251a1f39e0471f73b14ea1e60483ff19d59
Signed-off-by: Irina Mihai <irina.mihai@windriver.com>