Improve pylint checking in stx-config
Previously when pylint was invoked, all warnings were suppressed.
Now, the only warnings suppressed are the ones that are failing.
By making this change, fixed issues will not be reintroduced
by later commits.

Additional tasks will be created in the future to address the
currently suppressed warnings and errors.

This also adds a zuul job for pylint in sysinv.

Story: 2004515
Task: 28268
Change-Id: Ie162c08a9e70a70bfc089ebee711ef9d6a374f2b
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>