Docker Registry Can't Deploy On Computes
The flag marking the registry as "insecure" is missing on the computes.
This creates certificate issues and fails deployments on computes when
using images from the local Docker registry. This commit adds the
"insecure" flag to mark the registry as insecure on the compute nodes.
This will go away once proper authentication is implemented for
Docker registry.

Story: 2004520
Task: 28312

Change-Id: Ib8d9fad435eda85e1eaebc5fd71b0ff3d60f3dd8
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>