docker images management
Currently, the images pulled from public registries are used
by helm charts for pods launching. This commit updates this
behavior to pull images from the local docker registry to
launch pods. In order to pull images from local registry,
images for each supported chart are updated in sysinv overrides
to point to the local docker registry. For the non stx images
or images not included in sysinv overrides, they are handled
when application apply by getting them from charts and adding
the converted image tags to the override files.

A list of docker images required for stx-openstack app installation
is maintained in a file. When applying stx-openstack app (before
launching pods), if the required images do not exist in the local
registry, images will be pulled from the public registries and
pushed to the local registry.

To be able to use the local docker registry, the default disk
size for docker-distribution is increased from 1G to 16G.

Validated on AIO-SX/AIO-DX/2+2+2:
 - config_controller with kubernetes
 - During the initial stx-openstack app apply, images pulled from
   public registry and pushed to local registry. Images are pulled
   from local registry for pods launching on each node
 - Cleanup docker cache for each node, all images pulled from local
   registry during the subsequent application apply
 - Cleanup the images in local registry, images pulled from public
   and pushed to the local during the application apply

Story: 2004520
Task: 28526
Depends-On: https://review.openstack.org/#/c/626394/
Change-Id: I2001fff237cc2ccf4be1a8b15ab346d730bb32e9
Signed-off-by: Angie Wang <angie.wang@windriver.com>